+++ instrumentation/servlet/servlet-2.3/src/main/java/io/opentelemetry/instrumentation/hypertrace/servlet/v2_3/BufferingHttpServletRequest.java
@@ -29,6 +29,7 @@ import javax.servlet.ServletInputStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletRequestWrapper;
 import javax.servlet.http.HttpServletResponse;
+import org.hypertrace.agent.servlet.common.BufferedReaderWrapper;
 import org.hypertrace.agent.servlet.common.ByteBufferData;
 import org.hypertrace.agent.servlet.common.CharBufferData;
 import org.slf4j.Logger;
@@ -174,7 +175,7 @@ public class BufferingHttpServletRequest extends HttpServletRequestWrapper {
     if (reader == null) {
       reader = super.getReader();
       if (shouldReadContent()) {
-        reader = new BufferedReaderWrapper(reader, this);
+        reader = new BufferedReaderWrapper(reader, this.getCharBuffer());
       }
     }
     return reader;
@@ -334,124 +335,6 @@ public class BufferingHttpServletRequest extends HttpServletRequestWrapper {
     @Override
     public boolean markSupported() {
       return this.is.markSupported();
-    }
-  }
-
-  public static class BufferedReaderWrapper extends BufferedReader {
-
-    private static final Logger logger = LoggerFactory.getLogger(BufferedReaderWrapper.class);
-
-    private final BufferedReader reader;
-    private final BufferingHttpServletRequest bufferingHttpServletRequest;
-
-    public BufferedReaderWrapper(
-        BufferedReader reader, BufferingHttpServletRequest bufferingHttpServletRequest) {
-      super(reader);
-      this.reader = reader;
-      this.bufferingHttpServletRequest = bufferingHttpServletRequest;
-    }
-
-    @Override
-    public int read() throws IOException {
-      int read = this.reader.read();
-      try {
-        if (read >= 0) {
-          bufferingHttpServletRequest.getCharBuffer().appendData(read);
-        }
-      } catch (Exception e) {
-        logger.error("Error in read() - ", e);
-      }
-      return read;
-    }
-
-    @Override
-    public int read(char[] cbuf, int off, int len) throws IOException {
-      int read = this.reader.read(cbuf, off, len);
-      try {
-        if (read > 0) {
-          bufferingHttpServletRequest.getCharBuffer().appendData(cbuf, off, off + read);
-        }
-      } catch (Exception e) {
-        logger.error("Error in read(char[] cbuf, int off, int len) - ", e);
-      }
-      return read;
-    }
-
-    @Override
-    public String readLine() throws IOException {
-      String read = this.reader.readLine();
-      if (read == null) {
-        return null;
-      } else {
-        try {
-          CharBufferData charBufferData = bufferingHttpServletRequest.getCharBuffer();
-          charBufferData.appendData(read);
-        } catch (Exception e) {
-          logger.error("Error in readLine() - ", e);
-        }
-        return read;
-      }
-    }
-
-    @Override
-    public long skip(long n) throws IOException {
-      return this.reader.skip(n);
-    }
-
-    @Override
-    public boolean ready() throws IOException {
-      return this.reader.ready();
-    }
-
-    @Override
-    public boolean markSupported() {
-      return this.reader.markSupported();
-    }
-
-    @Override
-    public void mark(int readAheadLimit) throws IOException {
-      this.reader.mark(readAheadLimit);
-    }
-
-    @Override
-    public void reset() throws IOException {
-      this.reader.reset();
-    }
-
-    @Override
-    public void close() throws IOException {
-      this.reader.close();
-    }
-
-    @Override
-    public int read(java.nio.CharBuffer target) throws IOException {
-      int initPos = target.position();
-      int read = this.reader.read(target);
-      try {
-        if (read > 0) {
-          CharBufferData charBufferData = bufferingHttpServletRequest.getCharBuffer();
-
-          for (int i = initPos; i < initPos + read; ++i) {
-            charBufferData.appendData(target.get(i));
-          }
-        }
-      } catch (Exception e) {
-        logger.error("Error in read(target) - ", e);
-      }
-      return read;
-    }
-
-    @Override
-    public int read(char[] cbuf) throws IOException {
-      int read = this.reader.read(cbuf);
-      try {
-        if (read > 0) {
-          bufferingHttpServletRequest.getCharBuffer().appendData(cbuf, 0, read);
-        }
-      } catch (Exception e) {
-        logger.error("Error in read(char[]) - ", e);
-      }
-      return read;
     }
   }
 }
