+++ instrumentation/servlet/servlet-2.3/src/main/java/io/opentelemetry/instrumentation/hypertrace/servlet/v2_3/Servlet2BodyInstrumentation.java
@@ -100,13 +100,13 @@ public class Servlet2BodyInstrumentation extends Instrumenter.Default {
       "org.hypertrace.agent.core.DynamicConfig",
       "org.hypertrace.agent.servlet.common.ByteBufferData",
       "org.hypertrace.agent.servlet.common.CharBufferData",
+      "org.hypertrace.agent.servlet.common.BufferedWriterWrapper",
+      "org.hypertrace.agent.servlet.common.BufferedReaderWrapper",
       packageName + ".InstrumentationName",
       packageName + ".BufferingHttpServletResponse",
       packageName + ".BufferingHttpServletResponse$BufferingServletOutputStream",
-      packageName + ".BufferingHttpServletResponse$BufferedWriterWrapper",
       packageName + ".BufferingHttpServletRequest",
       packageName + ".BufferingHttpServletRequest$ServletInputStreamWrapper",
-      packageName + ".BufferingHttpServletRequest$BufferedReaderWrapper",
     };
   }
